feat: created the TransactionController Get Endpoints

+++ Constants/TransactionFeeTable.cs
@@ -4,12 +4,12 @@ namespace WalletAPI.Constants
 {
     public static class TransactionFeeTable
     {
-        public static readonly Dictionary<TransactionType, double> FeeRates = new()
+        public static readonly Dictionary<TransactionType, decimal> FeeRates = new()
         {
-            {TransactionType.Deposit, 0.0 },
-            {TransactionType.Withdraw, 0.015 },
-            {TransactionType.Transfer, 0.02 },
-            {TransactionType.Refund, 0.0 },
+            {TransactionType.Deposit, 0.0m },
+            {TransactionType.Withdraw, 0.015m },
+            {TransactionType.Transfer, 0.02m },
+            {TransactionType.Refund, 0.0m },
         };
     }
 }
