+++ Models/Wallet.cs
@@ -3,7 +3,7 @@
     public class Wallet
     {
         public int Id { get; set; }
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
         public DateTime CreatedAt { get; set; }
         public bool Active { get; set; }
         public int UserId { get; set; }
@@ -22,7 +22,7 @@
             UserId = user.Id;
             CreatedAt = DateTime.Now;
             Active = true;
-            Balance = 0.0;
+            Balance = 0.0m;
         }
 
         public bool IsActive()
